move to dce.py
